requirements.txt
# Newssys 2.0 - Python 依赖包

# ============================================================================
# 核心框架
# ============================================================================
fastapi==0.110.0
uvicorn[standard]==0.28.0
pydantic==2.7.0
pydantic-settings==2.3.0
python-dotenv==1.0.0

# ============================================================================
# 数据库
# ============================================================================
sqlalchemy>=2.0.0
aiomysql==0.2.0
aiosqlite>=0.19.0
pymysql==1.1.0
greenlet>=3.0.0

# ============================================================================
# HTTP 客户端
# ============================================================================
httpx==0.27.0
aiohttp==3.9.1

# ============================================================================
# HTML 解析
# ============================================================================
beautifulsoup4==4.12.3
lxml==5.1.0
html5lib==1.1

# ============================================================================
# 时间解析（非 AI）
# ============================================================================
python-dateutil==2.8.2
dateparser==1.2.0

# ============================================================================
# 文本处理
# ============================================================================
jieba==0.42.1

# ============================================================================
# 日志
# ============================================================================
python-json-logger==2.0.7

# ============================================================================
# 开发工具
# ============================================================================
pytest==8.1.1
pytest-asyncio==0.23.4
black==24.3.0
ruff==0.3.4
